JNC - refactoring of JNC examples - updated dependencies, minor fix in 0-intro Makefile - updated README and comments - history moved to CHANGES - example 1-interfaces - fix for new ssh connection
