+++ build.gradle
@@ -16,10 +16,10 @@ repositories {
 
 // Define dependencies
 dependencies {
-    implementation 'com.hierynomus:sshj:0.32.0'
-    implementation 'org.slf4j:slf4j-api:1.7.32'
+    implementation 'com.hierynomus:sshj:0.33.0'
+    implementation 'org.slf4j:slf4j-api:1.7.36'
     implementation 'org.bouncycastle:bcprov-jdk15on:1.70'
-    testImplementation 'junit:junit:4.11'
+    testImplementation 'junit:junit:4.13.2'
     testImplementation 'org.hamcrest:hamcrest-all:1.3'
 }
