--- doc/.gitignore
@@ -1,5 +1,0 @@
-*.aux
-*.d
-*.dvi
-*.log
-*.pdf
