--- doc/UserManual.tex
@@ -1,20 +1,0 @@
-\documentclass[a4paper,12pt]{memoir}
-\begin{document}
-
-This User Guide describes the JNC configuration management library. It provides
-an overview of the architecture of the library, its APIs including usage
-examples, and guidelines on how to extend the feature set.
-
-JNC also comes with a detailed API specification in Javadoc format.
-
-This guide is designed for developers who are planning to use the library. To
-use this document, you will need a broad understanding of the Java programming
-language, XML, YANG, and the NETCONF protocol.
-
-The examples in this guide requires access to a standards based NETCONF server.
-If we communicate with a NETCONF server with YANG support, we can utilize the
-data model at the server and compile a set of Java classes explicitly for the
-data model at the managed device, otherwise we must revert to a lower lever API
-that is not aware of the data model at the managed device.
-
-\end{document}
