+++ examples/0-intro/README
@@ -56,7 +56,7 @@ Source the rc script in the ConfD root directory:
 
   # source <confd_dir>/confdrc.tcsh
 
-Then change directory 
+Then change directory
 
   # cd <YOUR JNC DIR>/examples/0-intro
 
@@ -69,7 +69,7 @@ Build the ConfD node and start it:
 4.2 Building JNC example
 ========================
 
-Follow the instructions in the <YOUR JNC DIR>/README to install JNC.
+Follow the instructions in the <YOUR JNC DIR>/README to build JNC.
 
 Set your current working directory to the directory where
 this JNC example resides:
@@ -78,14 +78,14 @@ this JNC example resides:
 
 Build the JNC example:
 
-  # ant clean compile
+  # gradle clean build
 
 4.3 Run the Example
 ===============================
 
 Run this example by doing:
 
-  # ant run
+  # gradle run
 
 4.4 Stopping ConfD
 ===============================
@@ -101,7 +101,7 @@ After completing the examples above, stop ConfD as follows:
 
 You can also run the example above having ConfD and JNC on separate
 machines. If so, the following additional steps has to be made:
-This is the required procedure for users using the win32 version of 
+This is the required procedure for users using the win32 version of
 JNC.
 
 5.1 Building and Starting ConfD
@@ -132,4 +132,4 @@ or
 
   dev = new Device("mydev", duser, "192.168.1.1", 2022);
 
-Then rerun the ant script.
+Then rebuild.
