--- examples/0-intro/build.xml
@@ -1,88 +1,0 @@
-<project name="simple" basedir="." default="main">
-  <property environment="env"/>
-  <property name="src.dir"       value="src"/>
-  <property name="gen.dir"       value="src/gen"/>
-  <property name="build.dir"     value="build"/>
-  <property name="classes.dir"   value="${build.dir}/classes"/>
-  <property name="jar.dir"       value="${build.dir}/jar"/>
-  <property name="main-class"    value="dhcp.Main"/>
-  <property name="pyang"         value="pyang"/>
-  <property name="ganymed.dir"   value="../../jnc/lib"/>
-  <property name="jnc.dir"        value="../../jnc/build/jar"/>
-
-  <target name="clean">
-    <delete dir="${build.dir}"/>
-    <delete dir="${gen.dir}"/>
-    <delete file="Simple.schema"/>
-  </target>
-
-  <target name="generate">
-    <mkdir dir="${gen.dir}"/>
-    <mkdir dir="${gen.dir}/com"/>
-    <mkdir dir="${gen.dir}/com/example"/>
-    <mkdir dir="${gen.dir}/com/example/test"/>
-    <mkdir dir="${gen.dir}/com/example/test/inet"/>
-    <exec executable="${pyang}" failonerror="true">
-     <arg line="-f jnc --jnc-output src/gen/com/example/test ../yang/hosts.yang"/>
-    </exec>
-    <exec executable="${pyang}" failonerror="true">
-     <arg line="-f jnc --jnc-output src/gen/com/example/test/inet ${env.CONFD_DIR}/src/confd/yang/ietf-inet-types.yang"/>
-    </exec>
-  </target>
-
-  <target name="compile" depends="generate">
-    <mkdir dir="${classes.dir}"/>
-    <javac srcdir="${src.dir}" destdir="${classes.dir}" debug="on"
-           debuglevel="lines,vars,source" includeantruntime="false">
-      <classpath location="${jnc.dir}/JNC.jar"/>
-    </javac>
-  </target>
-
-  <target name="jar" depends="compile">
-    <mkdir dir="${jar.dir}"/>
-    <jar destfile="${jar.dir}/${ant.project.name}.jar" basedir="${classes.dir}">
-      <manifest>
-        <attribute name="Main-Class" value="${main-class}"/>
-      </manifest>
-    </jar>
-  </target>
-
-  <target name="run1" depends="compile">
-    <copy file="src/gen/com/example/test/Simple.schema" todir="."/>
-    <java classname="simple.Main" fork="true">
-      <arg value="-n"/>
-      <arg value="6"/>
-      <arg value="-t"/>
-      <classpath location="${jnc.dir}/JNC.jar"/>
-      <classpath location="${ganymed.dir}/ganymed-ssh2-262.jar"/>
-      <classpath location="build/classes" />
-    </java>
-  </target>
-
-  <target name="run" depends="compile" >
-    <copy file="src/gen/com/example/test/hosts/Simple.schema" todir="."/>
-    <java classname="simple.Main" fork="true">
-      <classpath location="${jnc.dir}/JNC.jar"/>
-      <classpath location="${ganymed.dir}/ganymed-ssh2-262.jar"/>
-      <classpath location="build/classes"/>
-    </java>
-  </target>
-
-  <target name="clean-build" depends="clean,jar"/>
-
-  <target name="main" depends="environ,compile,run"/>
-
-  <target name="notpresent" if="environ.badpath">
-    <fail message="** You need to set environment varible to point to JNC.jar"/>
-  </target>
-
-  <target name="environ">
-    <condition property="environ.badpath" >
-      <not>
-        <available classname="YangType"
-                   classpath="${jnc.dir}/JNC.jar"/>
-      </not>
-    </condition>
-    <antcall target="notpresent"/>
-  </target>
-</project>
