--- examples/1-interfaces/build.xml
@@ -1,56 +1,0 @@
-<project name="ietf-interfaces" basedir="." default="main">
-  <property environment="env" />
-  <property name="src.dir"     value="src" />
-  <property name="gen.dir"     value="src/gen" />
-  <property name="main-class"  value="ietf-interfaces.Client" />
-  <property name="build.dir"   value="build" />
-  <property name="classes.dir" value="${build.dir}/classes" />
-  <property name="jar.dir"     value="${build.dir}/jar" />
-  <property name="pyang"       value="pyang" />
-  <property name="jnc.dir"     value="../../jnc/build/jar" />
-  <property name="ganymed.dir" value="../../jnc/lib" />
-
-  <target name="clean">
-    <delete dir="${build.dir}" />
-    <delete dir="${gen.dir}" />
-    <delete file="ietf-interfaces.schema" />
-  </target>
-
-  <target name="generate">
-    <mkdir dir="${gen.dir}" />
-    <exec executable="${pyang}" failonerror="true">
-      <arg line="-f jnc --jnc-output ${gen.dir} ietf-interfaces.yang" />
-    </exec>
-  </target>
-
-  <target name="compile" depends="generate">
-    <mkdir dir="${classes.dir}" />
-    <javac srcdir="${src.dir}" destdir="${classes.dir}" debug="on"
-           debuglevel="lines,vars,source" includeantruntime="false">
-      <classpath location="${jnc.dir}/JNC.jar" />
-    </javac>
-  </target>
-
-  <target name="jar" depends="compile">
-    <jar destfile="${jar.dir}/${ant.project.name}.jar" basedir="${classes.dir}">
-      <manifest>
-        <attribute name="Main-Class" value="${main-class}" />
-      </manifest>
-    </jar>
-  </target>
-
-  <target name="Client"
-          depends="compile">
-    <java classname="app.Client"
-          failonerror="true"
-          fork="true">
-      <classpath location="${jnc.dir}/JNC.jar" />
-      <classpath location="${ganymed.dir}/ganymed-ssh2-262.jar" />
-      <classpath location="build/classes"/>
-    </java>
-  </target>
-
-  <target name="all" depends="jar" />
-  
-  <target name="main" depends="compile" />
-</project>
