+++ gradle.properties
@@ -1,4 +1,4 @@
 group                  =  com.tailf.jnc
 archivesBaseName       =  project.name
-version                =  1.0
+version                =  1.0.1
 author                 =  Tail-f Systems
