--- jnc/pom.xml
@@ -1,47 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <groupId>com.tailf</groupId>
-  <artifactId>jnc</artifactId>
-  <version>1.0</version>
-  <packaging>jar</packaging>
-
-  <name>JNC</name>
-  <description>JNC (Java NETCONF Client) is the name of a Java library for communicating with NETCONF agents.</description>
-  <url>https://github.com/tail-f-systems/JNC</url>
-
-  <licenses>
-    <license>
-      <name>Apache License, Version 2.0</name>
-      <url>http://apache.org/licenses/LICENSE-2.0</url>
-      <distribution>repo</distribution>
-    </license>
-  </licenses>
-
-  <developers>
-    <developer>
-      <organization>Tail-f Systems</organization>
-      <organizationUrl>http://www.tail-f.com/</organizationUrl>
-    </developer>
-  </developers>
-
-  <scm>
-    <connection>scm:git:git@github.com:tail-f-systems/JNC.git</connection>
-    <developerConnection>scm:git:git@github.com:tail-f-systems/JNC.git</developerConnection>
-    <url>git@github.com:tail-f-systems/JNC.git</url>
-  </scm>
-
-	<build>
-		<!-- Specify maven non-standard source folder -->
-		<sourceDirectory>src</sourceDirectory>
-	</build>
-	
-  <dependencies>
-    <dependency>
-      <groupId>ch.ethz.ganymed</groupId>
-      <artifactId>ganymed-ssh2</artifactId>
-      <version>262</version>
-    </dependency>
-  </dependencies>
-
-</project>
